requirements.txt
flask==3.1.0
flask-login==0.6.3
sqlalchemy==2.0.40
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
werkzeug==3.1.3
